+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - New Event: Chain sync lost notification (@Primexz)
+- New event: multi-channel backup notification. Backup is uploaded as file if ntfy is used as provider (@Feelancer21)
+
 ### Fixed
 ### Changed
 ### Removed